Added command to search for OID number based on volname
Finding out the hard way showed that lefthand reorganises the id number should a volume be removed, this therefore renders the original method of checking useless. The script now searches for the id using the volume name. The script now requires a -V followed by the volume name exactly as it is called in CMC, this option replaces the -O option.